fixes, docstrings
