+++ .eslintrc.cjs
@@ -5,10 +5,16 @@ module.exports = {
     "prettier",
   ],
   parser: "@typescript-eslint/parser",
-  plugins: ["@typescript-eslint"],
+  plugins: ["@typescript-eslint", "eslint-plugin-tsdoc"],
   root: true,
   parserOptions: {
     ecmaVersion: 2018,
     sourceType: "module",
+  },
+  env: {
+    node: true,
+  },
+  rules: {
+    "tsdoc/syntax": "warn",
   },
 }
