+++ .gitignore
@@ -60,4 +60,4 @@ typings/
 # next.js build output
 .next
 
-/dist
+/lib
