+++ tsconfig.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "declaration": true,
     "declarationMap": true,
-    "outDir": "./dist/",
+    "outDir": "./lib/",
     "sourceMap": true,
     "module": "commonjs",
     "target": "es2017",
